move compute_bin_time to spatial utils
