+++ spiketools/spatial/occupancy.py
@@ -7,6 +7,7 @@ import pandas as pd
 
 from spiketools.utils.checks import check_param_options, check_bin_range
 from spiketools.spatial.checks import check_position, check_position_bins
+from spiketools.spatial.utils import compute_bin_time
 
 ###################################################################################################
 ###################################################################################################
@@ -264,31 +265,6 @@ def normalize_bin_firing(bin_firing, occupancy):
         normalized_bin_firing = bin_firing / occupancy
 
     return normalized_bin_firing
-
-
-def compute_bin_time(timestamps):
-    """Compute the time duration of each position sample.
-
-    Parameters
-    ----------
-    timestamps : 1d array
-        Timestamps.
-
-    Returns
-    -------
-    1d array
-        Width, in time, of each bin.
-
-    Examples
-    --------
-    Compute times between timestamp samples:
-
-    >>> timestamp = np.array([0, 1.0, 3.0, 6.0, 8.0, 9.0])
-    >>> compute_bin_time(timestamp)
-    array([1., 2., 3., 2., 1., 0.])
-    """
-
-    return np.append(np.diff(timestamps), 0)
 
 
 def compute_occupancy(position, timestamps, bins, speed=None, speed_thresh=None, minimum=None,
